requirements.txt
attrs==24.3.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
chromium==0.0.0
Deprecated==1.2.15
exceptiongroup==1.2.2
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-appengine-logging==1.5.0
google-cloud-audit-log==0.3.0
google-cloud-core==2.4.1
google-cloud-logging==3.11.3
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.13.1
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
opentelemetry-api==1.29.0
outcome==1.3.0.post0
packaging==24.2
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==1.26.20
webdriver-manager==4.0.2
websocket-client==1.8.0
wrapt==1.17.0
wsproto==1.2.0
zipp==3.21.0
